// StageV/src/app/services/models/passport.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Passport {
  address?: string;
  dateOfBirth?: string;
  dateOfExpiry?: string;
  dateOfIssue?: string;
  firstName?: string;
  id?: number;
  issuingAuthority?: string;
  lastName?: string;
  nationality?: string;
  occupation?: string;
  placeOfBirth?: string;
  primaryPassport?: boolean;
  sex?: string;
}